Developed Color Picker
Had to design custom color picker, as kivy's color wheel is experimental and did not function properly with scatter layout or scroll view layout. Was not passing touches through to color picker properly.

This color picker uses sliders to determine a color, and provides a copy color button that applies the color the current layout (the base scatter layout).